Юзинги Почистил юзинги Переместил DTO в папку Data
