+++ Assets/Client/Scripts/Data/Features/Excursion.cs
@@ -1,7 +1,4 @@
-using System.Collections;
-using System.Collections.Generic;
 using Assets.Client.Scripts.Data.Features;
-using UnityEngine;
 
 public class Excursion: FeatureBase
 {
