+++ Assets/Client/Scripts/Data/Features/FeatureBase.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Assets.Client.Scripts.Data.Features
+﻿namespace Assets.Client.Scripts.Data.Features
 {
     public abstract class FeatureBase
     {
