+++ Assets/Client/Scripts/Data/Person.cs
@@ -1,0 +1,12 @@
+namespace Assets.Client.Scripts.Data
+{
+    public class Person
+    {
+        public string Name { get; set; }
+        public string Phrase { get; set; }
+        public string[] PreferredCountries { get; set; }
+        public string Budget { get; set; }
+        public string MaxRisc { get; set; }
+        public string Image { get; set; }
+    }
+}
