+++ Assets/Client/Scripts/Data/Tour.cs
@@ -1,0 +1,13 @@
+namespace Assets.Client.Scripts.Data
+{
+    /// <summary>
+    /// Сконфигурированный тур
+    /// </summary>
+    public class Tour
+    {
+        public Country Country;
+        public Transfer Transfer;
+        public Excursion Excursion;
+        public Hotel Hotel;
+    }
+}
