+++ Assets/Client/Scripts/Enums/EMessageType.cs
@@ -1,7 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
 public enum EMessageType
 {
     AboutMessage,
