+++ Assets/Client/Scripts/GUI/CountryListController.cs
@@ -1,7 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
 public class CountryListController : ListController
 {
     // Start is called before the first frame update
