+++ Assets/Client/Scripts/GUI/EventHandlers/AmoutOfMoneyHandler.cs
@@ -1,6 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
-using System.Net.Mime;
 using UnityEngine;
 using UnityEngine.UI;
