+++ Assets/Client/Scripts/GUI/EventHandlers/BackButtonHandler.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
