+++ Assets/Client/Scripts/GUI/EventHandlers/CommonBlockHandler.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 public class CommonBlockHandler : MonoBehaviour
