+++ Assets/Client/Scripts/GUI/EventHandlers/CountryControllerHandler.cs
@@ -1,8 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
-
 /// <summary>
 /// Класс обработки выбора страны
 /// </summary>
