+++ Assets/Client/Scripts/GUI/EventHandlers/ExcursionBlockHandler.cs
@@ -1,7 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
 /// <summary>
 /// Класс для обработки событий блока экскурсий
 /// </summary>
