+++ Assets/Client/Scripts/GUI/EventHandlers/ExcursionControllerHandler.cs
@@ -1,8 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
-
 public class ExcursionControllerHandler : ItemControllerHandler
 {
     protected override void OnBuyClickHandler()
