+++ Assets/Client/Scripts/GUI/EventHandlers/HotelBlockHandler.cs
@@ -1,7 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
 public class HotelBlockHandler : BlockHandler
 {
