+++ Assets/Client/Scripts/GUI/EventHandlers/HotelControllerHandler.cs
@@ -1,8 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
-
 public class HotelControllerHandler : ItemControllerHandler
 {
     protected override void OnBuyClickHandler()
