+++ Assets/Client/Scripts/GUI/ExcursionListController.cs
@@ -1,7 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
 /// <summary>
 /// Класс для обработки визуального списка экскурсий
 /// </summary>
