+++ Assets/Client/Scripts/GUI/HotelListController.cs
@@ -1,7 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
 public class HotelListController : ListController
 {
     public override void UpdateItems()
