+++ Assets/Client/Scripts/GUI/TransferListController.cs
@@ -1,7 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
 public class TransferListController : ListController
 {
     // Start is called before the first frame update
