--- Assets/Client/Scripts/Person.cs
@@ -1,13 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Person
-{
-    public string Name { get; set; }
-    public string Phrase { get; set; }
-    public string[] PreferredCountries { get; set; }
-    public string Budget { get; set; }
-    public string MaxRisc { get; set; }
-    public string Image { get; set; }
-}
