+++ Assets/Client/Scripts/Services/Implementations/ClientGenerator/OrderedClientGenerator.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using Assets.Client.Scripts.Data;
 using Assets.Client.Scripts.Services.Interfaces;
 
 namespace Assets.Client.Scripts.Services.Implementations.ClientGenerator
