+++ Assets/Client/Scripts/Services/Implementations/Loader/CountryLoader.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Xml;
+using Assets.Client.Scripts.Data;
 using Assets.Client.Scripts.Services.Implementations.Loader.Extensions;
 
 namespace Assets.Client.Scripts.Services.Implementations.Loader
