+++ Assets/Client/Scripts/Services/Interfaces/ITravelService.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Assets.Client.Scripts.Services.Interfaces
+﻿namespace Assets.Client.Scripts.Services.Interfaces
 {
     public interface ITravelService
     {
