--- Assets/Client/Scripts/Tour.cs
@@ -1,14 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-/// <summary>
-/// Сконфигурированный тур
-/// </summary>
-public class Tour
-{
-    public Country Country;
-    public Transfer Transfer;
-    public Excursion Excursion;
-    public Hotel Hotel;
-}
