+++ Assets/Client/Scripts/TourManager.cs
@@ -1,6 +1,5 @@
-using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
+using Assets.Client.Scripts.Data;
 using Assets.Client.Scripts.Services.Interfaces;
 using UnityEngine;
 using Zenject;
