+++ Assets/Client/Scripts/WorldLoader.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using Assets.Client.Scripts.Data;
 using Assets.Client.Scripts.Services.Interfaces;
 using UnityEngine;
 using Zenject;
